Separated utils classes to their respective areas of utility
