+++ noteworthyServiceLambda/src/main/java/com/nashss/se/noteworthy/services/s3/S3Utils.java
@@ -1,4 +1,4 @@
-package com.nashss.se.noteworthy.utils;
+package com.nashss.se.noteworthy.services.s3;
 
 public class S3Utils {
     public static final String INPUT_BUCKET = "nss-s3-c04-u7-noteworthy-cito.mcclure";
